Finish separating Spotify routes into its own Router file
